requirements.txt
pyttsx3
pydub
pillow
moviepy
imageio-ffmpeg
psutil



  